refactor: modularize components and partials
