+++ src/styles/main.scss
@@ -20,5 +20,5 @@ body {
     background: $light;
   }
 
-  @import './intro.scss';
+  @import './partials/intro';
 }
