+++ src/views/index.ejs
@@ -10,22 +10,6 @@
     <link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet" />
   </head>
   <body>
-    <div id="intro-wrapper" class="wrapper dark">
-      <div id="intro-container" class="container">
-        <div id="intro-video">
-          <video autoplay loop>
-            <source src="../assets/videos/banner-main-underlay.mp4" type="video/mp4" />
-          </video>
-        </div>
-        <div id="intro-info">
-          <div id="intro-photo-wrapper">
-            <div id="intro-photo"></div>
-          </div>
-          <div id="intro-name-wrapper">
-            <h1 id="intro-name">Anindya Kundu</h1>
-          </div>
-        </div>
-      </div>
-    </div>
+    <%- include partials/intro %>
   </body>
 </html>
